V1.4.2 更新日志(2016-11-15) 添加"重连有效时长"选项，限制重连的最大时间以避免手机过烫或android BLE工作栈出现异常
V1.4.1 更新日志(2016-11-14)
在Android6.0上测试的时候发现进行大量传输压力测试后断开连接会发现扫描不到设备(errorCode=SCAN_FAILED_APPLICATION_REGISTRATION_FAILED)。解决方案：在API>=21(即Android 5.0.1)时，重启蓝牙。
（注：Android 4.X是否存在该问题有待测试…… 由于不是必现，且没有4.X的手机，所以暂时未测）

V1.4.0 更新日志(2016-11-13)
添加文本框更新管理线程，周期性地检查是否需要更新文本(最短更新时间为1秒)
修改日志记录保留的条目数(1000 -> 200)
以上两个方法双管齐下，才得以修复由于频繁更新文本框内容导致的在“连续回传的压力测试”中界面假死问题

+++ app/src/main/java/com/newland/utils/PrefUtils.java
@@ -39,14 +39,14 @@ public class PrefUtils {
         return sp.getBoolean(key, defValue);
     }
 
-    public boolean putInt(String key, int value) {
+    public boolean putLong(String key, long value) {
         SharedPreferences.Editor editor = sp.edit();
-        editor.putInt(key, value);
+        editor.putLong(key, value);
         return editor.commit();
     }
 
-    public int getInt(String key) {
-        return sp.getInt(key, 0);
+    public long getLong(String key) {
+        return sp.getLong(key, 0);
     }
 
     public boolean putString(String key, String value) {
